[add] .env and sort recipe
