+++ package.json
@@ -16,6 +16,7 @@
     "axios": "^0.27.2",
     "babel-plugin-macros": "^3.1.0",
     "bootstrap": "^5.2.0",
+    "dotenv": "^16.0.2",
     "next": "12.2.3",
     "react": "18.2.0",
     "react-dom": "18.2.0",
