+++ pages/popularMenu.js
@@ -7,25 +7,31 @@ import Link from "next/link";
 function PopularMenu() {
   const [recipe, setRecipe] = useState([]);
 
+  const [filter, setFilter] = useState("DESC");
+
+  function handleAddrTypeChange(e) {
+    setFilter(e.target.value);
+  }
+
   React.useEffect(() => {
     axios
-      .get("http://localhost:8001/recipe")
+      .get(`${process.env.NEXT_PUBLIC_API}/recipe?filter=${filter}`)
       .then((res) => {
         setRecipe(res?.data?.recipe.slice(0, 6));
       })
       .catch((error) => {
         console.log(error?.response?.data);
       });
-  }, []);
+  }, [recipe]);
   return (
     <>
       <div className="container mt-4">
         <div className={popularStyle.title}>
           <div
-            className="d-flex align-content-center mb-4"
+            className="d-flex align-content-center mb-3"
             style={{ justifyContent: "space-between" }}
           >
-            <Link href="/home" passHref>
+            <Link href="/" passHref>
               <a>
                 <button type="button" className="btn btn-light">
                   <Image
@@ -40,6 +46,22 @@ function PopularMenu() {
             <h3 className="d-flex align-items-center">Popular Menu</h3>
             <div></div>
           </div>
+          <div className="d-flex justify-content-start mb-2">
+            <span>
+              <h5>Sort By:</h5>
+            </span>
+            <select
+              defaultValue={filter}
+              onChange={handleAddrTypeChange}
+              className="Default select example "
+              style={{ borderRadius: "10px", marginLeft: "5px" }}
+            >
+              <option selected value="DESC">
+                Newest
+              </option>
+              <option value="ASC">Latest</option>
+            </select>
+          </div>
           {recipe.map((item, key) => (
             <div key={key}>
               <Link
@@ -52,7 +74,7 @@ function PopularMenu() {
                     borderRadius: "15px",
                     padding: "10px",
                     border: "none",
-                    "box-shadow": "2px 2px 5px 1px rgba(0,0,0,0.12)",
+                    boxShadow: "2px 2px 5px 1px rgba(0,0,0,0.12)",
                     marginBottom: "20px",
                     cursor: "pointer",
                   }}
@@ -99,6 +121,38 @@ function PopularMenu() {
               </Link>
             </div>
           ))}
+          {/* pagination */}
+          {/* <div className="d-flex justify-content-center">
+            <nav aria-label="Page navigation example ">
+              <ul className="pagination">
+                <li className="page-item">
+                  <a className="page-link" href="#" aria-label="Previous">
+                    <span aria-hidden="true">&laquo;</span>
+                  </a>
+                </li>
+                <li className="page-item">
+                  <a className="page-link" href="#">
+                    1
+                  </a>
+                </li>
+                <li className="page-item">
+                  <a className="page-link" href="#">
+                    2
+                  </a>
+                </li>
+                <li className="page-item">
+                  <a className="page-link" href="#">
+                    3
+                  </a>
+                </li>
+                <li className="page-item">
+                  <a className="page-link" href="#" aria-label="Next">
+                    <span aria-hidden="true">&raquo;</span>
+                  </a>
+                </li>
+              </ul>
+            </nav>
+          </div> */}
         </div>
       </div>
     </>
